Refactor graphql hier
