+++ package.json
@@ -31,6 +31,7 @@
     "koa-cors": "0.0.16",
     "koa-graphql": "0.7.0",
     "koa-router": "7.2.1",
+    "lodash.merge": "4.6.0",
     "moment": "2.18.1",
     "request-promise": "4.2.1",
     "source-map-support": "0.4.15",
