+++ src/graphql/resolvers/index.ts
@@ -1,0 +1,5 @@
+import * as merge from 'lodash.merge';
+
+import jokes from './jokes';
+
+export default merge(jokes);
