+++ src/graphql/schema.graphqls
@@ -1,3 +1,9 @@
+# Root _Query_ definition
+type Query {
+  # Get _Joke_ by category
+  jokeByCategory(category: JokeCategoryEnum): Joke
+}
+
 # _Joke_ definition
 type Joke {
   id: String!
