+++ src/graphql/schema.ts
@@ -2,12 +2,12 @@ import * as fs from 'fs';
 import * as path from 'path';
 import { makeExecutableSchema } from 'graphql-tools';
 
-import { schema as jokesSchema, resolvers as jokesResolvers } from './jokes';
+import resolvers from './resolvers';
 
 
-const rootSchema: string = fs.readFileSync(path.join(__dirname, './query.graphqls'), 'utf8');
+const typeDefs: string = fs.readFileSync(path.join(__dirname, './schema.graphqls'), 'utf8');
 
 export default makeExecutableSchema({
-  typeDefs: [rootSchema, jokesSchema],
-  resolvers: { ...jokesResolvers },
+  typeDefs,
+  resolvers,
 });
