+++ src/graphql/types/jokes.ts
